refactor(settings): move section models into store domain
