+++ apps/desktop/src/stores/setting/sections/browser.ts
