+++ apps/desktop/src/stores/setting/sections/search.ts
