+++ apps/desktop/tests/SettingsView/settingsBrowserSection.test.ts
@@ -13,7 +13,7 @@ const browserSettingsPatch = vi.hoisted(() => ({
 const DEFAULT_BROWSER_DATA_PATH = 'C:/Users/test/AppData/Roaming/TouchAI/browser-data';
 
 vi.mock('@/stores/settings', async () => {
-    const { DEFAULT_BROWSER_SETTINGS } = await import('@/config/browserSettings');
+    const { DEFAULT_BROWSER_SETTINGS } = await import('@/stores/setting/sections/browser');
     const { ref } = await import('vue');
 
     return {
